[2단계 - 장바구니 기능] 메리(최승원) 미션 제출합니다.  (#313)
* feat: 서버 예외 처리 기능 구현

* refactor: 상품 삭제와 수정 요청 URI 통일

* refactor: Dao에서 Product 대신ProductEntity 사용, ProductEntity 빌더패턴 적용

* docs: 기능 명세 위치 이동 및 2단계 기능 명세 작성

* feat: 회원 테이블 추가

* feat: 설정 페이지 연결

* refactor: API 컨트롤러, 뷰 컨트롤러 분리

* docs: 장바구니 기능 명세 작성

* feat: 장바구니 페이지 연동 및 사용자 인증 후 상품 목록 반환

* feat: 장바구니 추가 API 구현

* feat: 장바구니 삭제 API 구현

* refactor: Rest컨트롤러 API admin 계층 삭제

* refactor: final 추가

* refactor: id 타입 통일

* refactor: 컨트롤러 분리

* refactor: 찾는 상품 없는 경우 예외 던지도록 수정

* refactor: Object Mapper 생성 및 Row mapper 분리

* fix: 상품 조회 예외 처리 수정

* feat: 카트에 존재하는 상품을 중복 추가하는 경우 예외 발생

* refactor: 카트에 상품 추가시 아이디 반환하도록 수정

* test: dao 테스트 추가

* test: api 컨트롤러 테스트 추가

* test: DB 조회 테스트 수정

* refactor: 로그인 정보 추출 방식 변경

* refactor: view 컨트롤러 통일

* refactor: API 변경

* fix: 잘못된 URI 연결 수정

* refactor: 서비스 -> 컨트롤러 의존성 분리

* feat: 상품 추가/삭제시 권한 체크

* refactor: 카트 도메인 생성

* test: 권한 체크를 위한 Authorization 추가

* refactor: 카트 추가 API 변경

* refactor: 카트 안에 여러 프로덕트 포함할 수 있도록 변경

* refactor: 장바구니 동일한 상품 삭제시 선택한 상품만 삭제하도록 변경

* refactor: ObjectMapper 삭제

* fix: 장바구니 저장시 member_id가 잘못 저장되는 오류 수정

* refactor: 레포지토리 계층 추가

* refactor: 불필요한 변수, 메소드 삭제

* refactor: 불필요한 변수, 메소드 삭제

* style: 안쓰는 주석 제거

+++ src/main/resources/static/js/admin.js
@@ -46,9 +46,19 @@ form.addEventListener('submit', (event) => {
 
 // TODO: [1단계] 상품 관리 CRUD API에 맞게 변경
 const createProduct = (product) => {
+    const credentials = localStorage.getItem('credentials');
+    if (!credentials) {
+        alert('사용자 정보가 없습니다.');
+        window.location.href = '/settings';
+        return;
+    }
+
     axios.request({
         method: 'post',
-        url: '/admin/product',
+        url: '/products',
+        headers: {
+            'Authorization': `Basic ${credentials}`
+        },
         data : product
     }).then((response) => {
         window.location.reload();
@@ -59,12 +69,22 @@ const createProduct = (product) => {
 
 // TODO: [1단계] 상품 관리 CRUD API에 맞게 변경
 const updateProduct = (product) => {
+    const credentials = localStorage.getItem('credentials');
+    if (!credentials) {
+        alert('사용자 정보가 없습니다.');
+        window.location.href = '/settings';
+        return;
+    }
+
     const { id } = product;
 
     axios.request({
         method: 'put',
-        url: '/admin/product',
-        data : product
+        url: '/products/' + id,
+        headers: {
+            'Authorization': `Basic ${credentials}`
+        },
+        data: product
     }).then((response) => {
         window.location.reload();
     }).catch((error) => {
@@ -74,9 +94,19 @@ const updateProduct = (product) => {
 
 // TODO: [1단계] 상품 관리 CRUD API에 맞게 변경
 const deleteProduct = (id) => {
+    const credentials = localStorage.getItem('credentials');
+    if (!credentials) {
+        alert('사용자 정보가 없습니다.');
+        window.location.href = '/settings';
+        return;
+    }
+
     axios.request({
         method: 'delete',
-        url: `/admin/product/${id}`
+        headers: {
+            'Authorization': `Basic ${credentials}`
+        },
+        url: '/products/' + id,
     }).then((response) => {
         window.location.reload();
     }).catch((error) => {
